yubikey-agent: drop transaction on SIGHUP
When running yubikey-agent has an exclusive lock on the YubiKey which
applies to both PIV and PGP applets (presumably because they are
mediated by the same daemon) but not FIDO2. This is necessary in order
to keep the PIN cached, see go-piv/piv-go#47.

Handle SIGHUP by releasing that lock. Never was a signal name more apt.

This allows using "killall -HUP yubikey-agent" in a script or alias to
release the applet when needed.

$ ssh -T git@github.com
$ killall -HUP yubikey-agent; pass filippo.io
$ killall gpg-agent # couldn't find a command to make it drop its tx

yubikey-agent will seamlessly make a new transaction and ask for a PIN
the next time a signing operation is requested.