systemd: remove options unsupported in user daemon
Many unit options are documented to not work in the user daemons,
usually with the following note in the systemd.exec(5) man page:

> This option is only available for system services and is not supported
> for services running in per-user instances of the service manager.

In some circumstances these settings work where user namespaces are
supported (by setting the `PrivateUser` option). However users
namespaces are disabled on most distributions.

This changeset pares the options to those supported in user daemons to
allow for ease of use for users.

Fixes: #49